🔨 consolidate wordpress types
